+++ pkg/codegen/go/gen.go
@@ -994,12 +994,6 @@ func (pkg *pkgContext) genConfig(w io.Writer, variables []*schema.Property) erro
 	return nil
 }
 
-func (pkg *pkgContext) genPackageRegistration(w io.Writer) {
-	fmt.Fprintf(w, "func init() {\n")
-	fmt.Fprintf(w, "\tpulumi.RegisterPackage(pulumi.PackageInfo{Name:\"%s\", Version:\"%s\"})\n", pkg.pkg.Name, pkg.pkg.Version.String())
-	fmt.Fprintf(w, "}\n")
-}
-
 func (pkg *pkgContext) getPkg(mod string) *pkgContext {
 	if override, ok := pkg.modToPkg[mod]; ok {
 		mod = override
@@ -1172,7 +1166,7 @@ func GeneratePackage(tool string, pkg *schema.Package) (map[string][]byte, error
 		files[relPath] = formattedSource
 	}
 
-	name, registerPackage := pkg.Name, pkg.Provider != nil
+	name := pkg.Name
 	for _, mod := range pkgMods {
 		pkg := packages[mod]
 
@@ -1247,16 +1241,6 @@ func GeneratePackage(tool string, pkg *schema.Package) (map[string][]byte, error
 			pkg.genTypeRegistrations(buffer, pkg.types)
 
 			setFile(path.Join(mod, "pulumiTypes.go"), buffer.String())
-		}
-
-		// Package registration
-		if registerPackage {
-			buffer := &bytes.Buffer{}
-
-			pkg.genHeader(buffer, []string{"github.com/pulumi/pulumi/sdk/go/pulumi"}, nil)
-			pkg.genPackageRegistration(buffer)
-
-			setFile(path.Join(mod, "pulumiManifest.go"), buffer.String())
 		}
 
 		// Utilities