+++ sdk/go/pulumi-language-go/main.go
@@ -15,20 +15,25 @@
 package main
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"flag"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"syscall"
 
+	"github.com/blang/semver"
 	pbempty "github.com/golang/protobuf/ptypes/empty"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 
+	"github.com/pulumi/pulumi/sdk/go/common/util/buildutil"
 	"github.com/pulumi/pulumi/sdk/go/common/util/cmdutil"
 	"github.com/pulumi/pulumi/sdk/go/common/util/logging"
 	"github.com/pulumi/pulumi/sdk/go/common/util/rpcutil"
@@ -157,10 +162,109 @@ func newLanguageHost(engineAddress, tracing string) pulumirpc.LanguageRuntimeSer
 	}
 }
 
+// modInfo is the useful portion of the output from `go list -m -json all`
+// with respect to plugin acquisition
+type modInfo struct {
+	Path    string
+	Version string
+}
+
+func (m *modInfo) getPlugin() (*pulumirpc.PluginDependency, error) {
+	if !strings.HasPrefix(m.Path, "github.com/pulumi/pulumi-") {
+		return nil, errors.New("module is not a pulumi provider")
+	}
+
+	// github.com/pulumi/pulumi-aws/sdk/... => aws
+	pluginPart := strings.Split(m.Path, "/")[2]
+	name := strings.SplitN(pluginPart, "-", 2)[1]
+
+	v, err := semver.ParseTolerant(m.Version)
+	if err != nil {
+		return nil, errors.New("module does not have semver compatible version")
+	}
+	version := m.Version
+
+	// psuedoversions are commits that don't have a corresponding tag at the specified git hash
+	// https://golang.org/cmd/go/#hdr-Pseudo_versions
+	// pulumi-aws v1.29.1-0.20200403140640-efb5e2a48a86 (first commit after 1.29.0 release)
+	if buildutil.IsPseudoVersion(version) {
+		// no prior tag means there was never a release build
+		if v.Major == 0 && v.Minor == 0 && v.Patch == 0 {
+			return nil, errors.New("invalid pseduoversion with no prior tag")
+		}
+		// patch is typically bumped from the previous tag when using pseudo version
+		// downgrade the patch by 1 to make sure we match a release that exists
+		patch := v.Patch
+		if patch > 0 {
+			patch--
+		}
+		version = fmt.Sprintf("v%v.%v.%v", v.Major, v.Minor, patch)
+	}
+
+	plugin := &pulumirpc.PluginDependency{
+		Name:    name,
+		Version: version,
+		Kind:    "resource",
+	}
+
+	return plugin, nil
+}
+
 // GetRequiredPlugins computes the complete set of anticipated plugins required by a program.
+// We're lenient here as this relies on the `go list` command and the use of modules.
+// If the consumer insists on using some other form of dependency management tool like
+// dep or glide, the list command fails with "go list -m: not using modules"
 func (host *goLanguageHost) GetRequiredPlugins(ctx context.Context,
 	req *pulumirpc.GetRequiredPluginsRequest) (*pulumirpc.GetRequiredPluginsResponse, error) {
-	return &pulumirpc.GetRequiredPluginsResponse{}, nil
+
+	logging.V(5).Infof("GetRequiredPlugins: Determining pulumi packages")
+
+	gobin, err := findExecutable("go")
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn't find go binary")
+	}
+
+	// don't wire up stderr so non-module users don't see error output from list
+	cmd := exec.Command(gobin, "list", "-m", "-json", "all")
+	cmd.Env = os.Environ()
+
+	stdout, err := cmd.Output()
+	if err != nil {
+		// will err if the project isn't using modules
+		logging.V(5).Infof("GetRequiredPlugins: Error discovering plugin requirements: %s", err.Error())
+		return &pulumirpc.GetRequiredPluginsResponse{}, nil
+	}
+
+	plugins := []*pulumirpc.PluginDependency{}
+
+	dec := json.NewDecoder(bytes.NewReader(stdout))
+	for {
+		var m modInfo
+		if err := dec.Decode(&m); err != nil {
+			if err == io.EOF {
+				break
+			}
+			logging.V(5).Infof("GetRequiredPlugins: Error parsing list output: %s", err.Error())
+			return &pulumirpc.GetRequiredPluginsResponse{}, nil
+		}
+
+		plugin, err := m.getPlugin()
+		if err == nil {
+			logging.V(5).Infof("GetRequiredPlugins: Found plugin name: %s, version: %s", plugin.Name, plugin.Version)
+			plugins = append(plugins, plugin)
+		} else {
+			logging.V(5).Infof(
+				"GetRequiredPlugins: Ignoring dependency: %s, version: %s, error: %s",
+				m.Path,
+				m.Version,
+				err.Error(),
+			)
+		}
+	}
+
+	return &pulumirpc.GetRequiredPluginsResponse{
+		Plugins: plugins,
+	}, nil
 }
 
 // RPC endpoint for LanguageRuntimeServer::Run